[Fix] Test App Dealing Wrong Commands
