+++ shell.py
@@ -1,10 +1,10 @@
-import re
 import sys
 
 from shell_command import ShellCommand
 from command_factory import ShellCommandFactory
 from logger import Logger
 from test_app import TestApp
+from utils import is_ssd_command
 
 
 class Shell:
@@ -30,7 +30,7 @@ class Shell:
     def parse_command(self, command: str) -> ShellCommand or TestApp:
         self.logger.log("command 파싱")
         command = command.strip()
-        if self.is_ssd_command(command):
+        if is_ssd_command(command):
             return ShellCommandFactory(command).create_command()
         else:
             try:
@@ -39,19 +39,6 @@ class Shell:
                 self.logger.log("파싱 실패 ValueError 발생")
                 raise ValueError
 
-    @staticmethod
-    def is_ssd_command(command):
-        if command in ["help", "exit", "fullread"]:
-            return True
-        if re.fullmatch(r"read [0-9]{1,2}", command):
-            return True
-        if re.fullmatch(r"write [0-9]{1,2} 0x[0-9A-F]{8}", command):
-            return True
-        if re.fullmatch(r"fullwrite 0x[0-9A-F]{8}", command):
-            return True
-        if re.fullmatch(r"(erase|erase_range) [0-9]{1,2} \b(?:100|\d{1,2})\b$", command):
-            return True
-        return False
 
     def run_scripts(self):
         for script in self.scripts:
